Change file permission for create script file
In newer version of MS SQL server the database engine is not run under root. When sqlcmd is invoked it returns an error code about permission errors. By setting the script file as ReadWrite for all parts 'chmod 666' sqlcmd is able to read the script file and create the database.