Helper methods for receiving a message
When waiting for a message to be received on a SubscriptionClient the logic can get complicated and redundant.

The class SubscriptionClientExtensions provides some convient methods that can be used to await and deserialize a message to a concrete DTO.

The inner workings are based on TaskCompletionSource<T> as shown in the examples on the website. Besides that a timeout options is added.